move HIDAliases.h
